% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_ana.R
\name{eda_single}
\alias{eda_single}
\title{Perform Exploratory Data Analysis on Time Series}
\usage{
eda_single(series, m = NULL, s = NULL)
}
\arguments{
\item{series}{Numerical vector containing NA for missing values.}

\item{m}{Optional, scaling factor (autocalculated).}

\item{s}{Optional, cutoff series length (autocalculated: default 6*m).}
}
\value{
A result DataFrame to pass to plotting_single().
}
\description{
Perform Exploratory Data Analysis on Time Series
}
