% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_kendall_cor.R
\name{kendall_cor_test}
\alias{kendall_cor_test}
\title{Kendall Correlation Test}
\usage{
kendall_cor_test(x, y, alternative = c("two.sided", "greater", "less"))
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{alternative}{Type of alternative hypothesis ("two.sided", "greater", "less")}
}
\value{
htest object with results
}
\description{
Perform significance test for Kendall's tau-a correlation
}
